# Improve encoding and decoding performance for native histograms in new internal query result payload format