--- a/pkg/api/protobuf_codec_test.go
+++ b/pkg/api/protobuf_codec_test.go
@@ -301,10 +301,9 @@ var protobufCodecScenarios = map[string]struct {
 				Vector: &mimirpb.VectorData{
 					Histograms: []mimirpb.VectorHistogram{
 						{
-							Metric: []string{"name-1", "value-1"},
+							Metric:                []string{"name-1", "value-1"},
+							TimestampMilliseconds: 1234,
 							Histogram: mimirpb.FloatHistogram{
-								Timestamp: 1234,
-
 								CounterResetHint: mimirpb.Histogram_GAUGE,
 								Schema:           3,
 								ZeroThreshold:    1.23,
@@ -405,10 +404,9 @@ var protobufCodecScenarios = map[string]struct {
 				Vector: &mimirpb.VectorData{
 					Histograms: []mimirpb.VectorHistogram{
 						{
-							Metric: []string{"name-1", "value-1"},
+							Metric:                []string{"name-1", "value-1"},
+							TimestampMilliseconds: 1234,
 							Histogram: mimirpb.FloatHistogram{
-								Timestamp: 1234,
-
 								CounterResetHint: mimirpb.Histogram_GAUGE,
 								Schema:           3,
 								ZeroThreshold:    1.23,
@@ -428,10 +426,9 @@ var protobufCodecScenarios = map[string]struct {
 							},
 						},
 						{
-							Metric: []string{"name-3", "value-3"},
+							Metric:                []string{"name-3", "value-3"},
+							TimestampMilliseconds: 12340,
 							Histogram: mimirpb.FloatHistogram{
-								Timestamp: 12340,
-
 								CounterResetHint: mimirpb.Histogram_GAUGE,
 								Schema:           4,
 								ZeroThreshold:    1.203,
@@ -740,26 +737,27 @@ var protobufCodecScenarios = map[string]struct {
 					Series: []mimirpb.MatrixSeries{
 						{
 							Metric: []string{"name-1", "value-1", "name-2", "value-2"},
-							Histograms: []mimirpb.FloatHistogram{
+							Histograms: []mimirpb.MatrixHistogram{
 								{
-									Timestamp: 1234,
-
-									CounterResetHint: mimirpb.Histogram_GAUGE,
-									Schema:           3,
-									ZeroThreshold:    1.23,
-									ZeroCount:        456,
-									Count:            9001,
-									Sum:              789.1,
-									PositiveSpans: []mimirpb.BucketSpan{
-										{Offset: 4, Length: 1},
-										{Offset: 3, Length: 2},
+									TimestampMilliseconds: 1234,
+									Histogram: mimirpb.FloatHistogram{
+										CounterResetHint: mimirpb.Histogram_GAUGE,
+										Schema:           3,
+										ZeroThreshold:    1.23,
+										ZeroCount:        456,
+										Count:            9001,
+										Sum:              789.1,
+										PositiveSpans: []mimirpb.BucketSpan{
+											{Offset: 4, Length: 1},
+											{Offset: 3, Length: 2},
+										},
+										NegativeSpans: []mimirpb.BucketSpan{
+											{Offset: 7, Length: 3},
+											{Offset: 9, Length: 1},
+										},
+										PositiveBuckets: []float64{100, 200, 300},
+										NegativeBuckets: []float64{400, 500, 600, 700},
 									},
-									NegativeSpans: []mimirpb.BucketSpan{
-										{Offset: 7, Length: 3},
-										{Offset: 9, Length: 1},
-									},
-									PositiveBuckets: []float64{100, 200, 300},
-									NegativeBuckets: []float64{400, 500, 600, 700},
 								},
 							},
 						},
@@ -849,46 +847,48 @@ var protobufCodecScenarios = map[string]struct {
 									Value:                 90.01,
 								},
 							},
-							Histograms: []mimirpb.FloatHistogram{
+							Histograms: []mimirpb.MatrixHistogram{
 								{
-									Timestamp: 1234,
-
-									CounterResetHint: mimirpb.Histogram_GAUGE,
-									Schema:           3,
-									ZeroThreshold:    1.23,
-									ZeroCount:        456,
-									Count:            9001,
-									Sum:              789.1,
-									PositiveSpans: []mimirpb.BucketSpan{
-										{Offset: 4, Length: 1},
-										{Offset: 3, Length: 2},
+									TimestampMilliseconds: 1234,
+									Histogram: mimirpb.FloatHistogram{
+										CounterResetHint: mimirpb.Histogram_GAUGE,
+										Schema:           3,
+										ZeroThreshold:    1.23,
+										ZeroCount:        456,
+										Count:            9001,
+										Sum:              789.1,
+										PositiveSpans: []mimirpb.BucketSpan{
+											{Offset: 4, Length: 1},
+											{Offset: 3, Length: 2},
+										},
+										NegativeSpans: []mimirpb.BucketSpan{
+											{Offset: 7, Length: 3},
+											{Offset: 9, Length: 1},
+										},
+										PositiveBuckets: []float64{100, 200, 300},
+										NegativeBuckets: []float64{400, 500, 600, 700},
 									},
-									NegativeSpans: []mimirpb.BucketSpan{
-										{Offset: 7, Length: 3},
-										{Offset: 9, Length: 1},
-									},
-									PositiveBuckets: []float64{100, 200, 300},
-									NegativeBuckets: []float64{400, 500, 600, 700},
 								},
 								{
-									Timestamp: 12340,
-
-									CounterResetHint: mimirpb.Histogram_GAUGE,
-									Schema:           4,
-									ZeroThreshold:    1.203,
-									ZeroCount:        4560,
-									Count:            90010,
-									Sum:              7890.1,
-									PositiveSpans: []mimirpb.BucketSpan{
-										{Offset: 40, Length: 1},
-										{Offset: 30, Length: 2},
+									TimestampMilliseconds: 12340,
+									Histogram: mimirpb.FloatHistogram{
+										CounterResetHint: mimirpb.Histogram_GAUGE,
+										Schema:           4,
+										ZeroThreshold:    1.203,
+										ZeroCount:        4560,
+										Count:            90010,
+										Sum:              7890.1,
+										PositiveSpans: []mimirpb.BucketSpan{
+											{Offset: 40, Length: 1},
+											{Offset: 30, Length: 2},
+										},
+										NegativeSpans: []mimirpb.BucketSpan{
+											{Offset: 70, Length: 3},
+											{Offset: 90, Length: 1},
+										},
+										PositiveBuckets: []float64{1000, 2000, 3000},
+										NegativeBuckets: []float64{4000, 5000, 6000, 7000},
 									},
-									NegativeSpans: []mimirpb.BucketSpan{
-										{Offset: 70, Length: 3},
-										{Offset: 90, Length: 1},
-									},
-									PositiveBuckets: []float64{1000, 2000, 3000},
-									NegativeBuckets: []float64{4000, 5000, 6000, 7000},
 								},
 							},
 						},