… pkg/mimirpb/mimir.pb.go