#### What this PR does

This PR improves the performance of both encoding and decoding native histograms in the new internal query result payload format.

<details><summary>Benchmark results: encoding in querier</summary>

```
name                                                                              old time/op    new time/op    delta
ProtobufCodec_Encode/vector_with_series_with_histogram_value-10                      416ns ± 1%     349ns ± 0%  -16.03%  (p=0.008 n=5+5)
ProtobufCodec_Encode/vector_with_float_and_histogram_values-10                       837ns ± 1%     716ns ± 0%  -14.40%  (p=0.008 n=5+5)
ProtobufCodec_Encode/matrix_with_series_with_histogram_value-10                      466ns ± 2%     426ns ± 0%   -8.46%  (p=0.008 n=5+5)
ProtobufCodec_Encode/matrix_with_series_with_both_float_and_histogram_values-10      716ns ± 0%     649ns ± 1%   -9.32%  (p=0.016 n=4+5)

name                                                                              old alloc/op   new alloc/op   delta
ProtobufCodec_Encode/vector_with_series_with_histogram_value-10                       456B ± 0%      424B ± 0%   -7.02%  (p=0.008 n=5+5)
ProtobufCodec_Encode/vector_with_float_and_histogram_values-10                      1.06kB ± 0%    1.00kB ± 0%   -6.02%  (p=0.008 n=5+5)
ProtobufCodec_Encode/matrix_with_series_with_histogram_value-10                       544B ± 0%      496B ± 0%   -8.82%  (p=0.008 n=5+5)
ProtobufCodec_Encode/matrix_with_series_with_both_float_and_histogram_values-10       944B ± 0%      848B ± 0%  -10.17%  (p=0.008 n=5+5)

name                                                                              old allocs/op  new allocs/op  delta
ProtobufCodec_Encode/vector_with_series_with_histogram_value-10                       7.00 ± 0%      5.00 ± 0%  -28.57%  (p=0.008 n=5+5)
ProtobufCodec_Encode/vector_with_float_and_histogram_values-10                        13.0 ± 0%       9.0 ± 0%  -30.77%  (p=0.008 n=5+5)
ProtobufCodec_Encode/matrix_with_series_with_histogram_value-10                       8.00 ± 0%      6.00 ± 0%  -25.00%  (p=0.008 n=5+5)
ProtobufCodec_Encode/matrix_with_series_with_both_float_and_histogram_values-10       11.0 ± 0%       7.0 ± 0%  -36.36%  (p=0.008 n=5+5)
```

</details>

<details><summary>Benchmark results: decoding in query-frontend</summary>

```
name                                                                                                            old time/op    new time/op    delta
ProtobufFormat_DecodeResponse/successful_vector_response_with_histogram_value-10                                  2.56µs ± 0%    2.51µs ± 1%  -1.80%  (p=0.008 n=5+5)
ProtobufFormat_DecodeResponse/successful_vector_response_with_float_and_histogram_values-10                       2.77µs ± 0%    2.75µs ± 0%  -0.46%  (p=0.040 n=5+5)
ProtobufFormat_DecodeResponse/successful_matrix_response_with_histogram_value-10                                  2.73µs ± 1%    2.67µs ± 0%  -2.49%  (p=0.008 n=5+5)
ProtobufFormat_DecodeResponse/successful_matrix_response_with_float_and_histogram_values-10                       2.76µs ± 1%    2.67µs ± 1%  -3.14%  (p=0.008 n=5+5)

name                                                                                                            old alloc/op   new alloc/op   delta
ProtobufFormat_DecodeResponse/successful_vector_response_with_histogram_value-10                                  2.97kB ± 0%    2.94kB ± 0%  -1.08%  (p=0.008 n=5+5)
ProtobufFormat_DecodeResponse/successful_vector_response_with_float_and_histogram_values-10                       3.18kB ± 0%    3.14kB ± 0%  -1.01%  (p=0.008 n=5+5)
ProtobufFormat_DecodeResponse/successful_matrix_response_with_histogram_value-10                                  3.12kB ± 0%    3.07kB ± 0%  -1.54%  (p=0.008 n=5+5)
ProtobufFormat_DecodeResponse/successful_matrix_response_with_float_and_histogram_values-10                       3.15kB ± 0%    3.10kB ± 0%  -1.52%  (p=0.008 n=5+5)

name                                                                                                            old allocs/op  new allocs/op  delta
ProtobufFormat_DecodeResponse/successful_vector_response_with_histogram_value-10                                    50.0 ± 0%      48.0 ± 0%  -4.00%  (p=0.008 n=5+5)
ProtobufFormat_DecodeResponse/successful_vector_response_with_float_and_histogram_values-10                         57.0 ± 0%      55.0 ± 0%  -3.51%  (p=0.008 n=5+5)
ProtobufFormat_DecodeResponse/successful_matrix_response_with_histogram_value-10                                    54.0 ± 0%      52.0 ± 0%  -3.70%  (p=0.008 n=5+5)
ProtobufFormat_DecodeResponse/successful_matrix_response_with_float_and_histogram_values-10                         56.0 ± 0%      54.0 ± 0%  -3.57%  (p=0.008 n=5+5)
```

</details>


Note that it requires a breaking change to the protobuf schema, but I believe this is worthwhile given native histograms are an experimental feature and we're not yet running it in production.

#### Which issue(s) this PR fixes or relates to

#4104

#### Checklist

- [x] Tests updated
- [n/a] Documentation added
- [ ] `CHANGELOG.md` updated - the order of entries should be `[CHANGE]`, `[FEATURE]`, `[ENHANCEMENT]`, `[BUGFIX]`